+++ build/master/copy-artifacts.sh
@@ -15,9 +15,13 @@ export ARTIFACT_DIR=/var/jbpm-artifacts/master
 
  #cleanup of $ARTIFACT_DIR/new where all built artifacts will be copied to
 
-rm -rf $ARTIFACT_DIR/new
+if [ -d "$ARTIFACT_DIR"/new ]; then
+   rm -rf $ARTIFACT_DIR/new
+fi
+
 mkdir $ARTIFACT_DIR/new
 mkdir $ARTIFACT_DIR/new/docs
+mkdir $ARTIFACT_DIR/new/examples
 
 # copy all artifacts to $ARTIFACT_DIR/new
 
@@ -54,7 +58,6 @@ cp $GIT_DIR/jbpm/jbpm-distribution/target/jbpm-*-bin.zip $ARTIFACT_DIR/new
 # org.drools.updatesite
 cp $GIT_DIR/droolsjbpm-tools/drools-eclipse/org.drools.updatesite/target/org.drools.updatesite-*.zip $ARTIFACT_DIR/new
 # drools examples
-mkdir $ARTIFACT_DIR/new/examples
 cp -r $GIT_DIR/droolsjbpm-build-distribution/droolsjbpm-uber-distribution/target/droolsjbpm-uber-distribution-*/downloads_htdocs/drools/release/*/*.zip $ARTIFACT_DIR/new/examples
 cp -r $GIT_DIR/droolsjbpm-build-distribution/droolsjbpm-uber-distribution/target/droolsjbpm-uber-distribution-*/downloads_htdocs/optaplanner/release/*/*.zip $ARTIFACT_DIR/new/examples
 # kie-server*.war
